fix(box2d): eliminate FMA fusion so results match on amd64 and arm64
The golden bit-pattern tests were generated on arm64 and failed on the
CI amd64 runner. Root cause: the package's FMA rule only guarded
products written inline as an operand of + or -. Fusion is also legal
when an unrounded product reaches the + through a local, a struct
field, or a function parameter that is later inlined -- e.g.
MulSV's `s * v.X` flowing into mulAdd's addend, which arm64 emitted as
FMADDD and baseline amd64 did not. Disassembly showed 161 such sites.

Round products where they are formed, and additionally round the
operands the math_fma.go helpers receive so a caller's product is
rounded once the call is inlined (the smaller fix at most sites).
arm64 now emits zero FMA instructions, and the regenerated goldens
carry exactly the values amd64 was already producing.

Add TestNoFusedMultiplyAdd, which compiles the package for arm64 and
amd64/v3 and fails on any FMA instruction in the assembly. Reviewing
source for this is unreliable, so the invariant is now machine-checked
and wired into the determinism workflow ahead of the golden tests, so
a regression reports its cause rather than just "bits differ".

+++ pkg/box2d/contact_solver.go
@@ -404,7 +404,7 @@ func (w *World) solveContactsColor(ctx *stepContext, colorIndex int, useBias boo
 			vt := Dot(Sub(vrB, vrA), tangent) - constraint.tangentSpeed
 
 			// incremental tangent impulse
-			impulse := cp.tangentMass * -vt
+			impulse := float64(cp.tangentMass * -vt)
 
 			// clamp the accumulated force
 			maxFriction := friction * cp.normalImpulse
@@ -422,7 +422,7 @@ func (w *World) solveContactsColor(ctx *stepContext, colorIndex int, useBias boo
 
 		// Rolling resistance
 		{
-			deltaLambda := -constraint.rollingMass * (wB - wA)
+			deltaLambda := float64(-constraint.rollingMass * (wB - wA))
 			lambda := constraint.rollingImpulse
 			maxLambda := constraint.rollingResistance * totalNormalImpulse
 			constraint.rollingImpulse = clampFloat(lambda+deltaLambda, -maxLambda, maxLambda)
@@ -517,7 +517,7 @@ func (w *World) applyRestitutionColor(ctx *stepContext, colorIndex int) {
 
 			// compute normal impulse
 			// impulse = -normalMass * (vn + restitution * relativeVelocity)
-			impulse := -cp.normalMass * (vn + float64(restitution*cp.relativeVelocity))
+			impulse := float64(-cp.normalMass * (vn + float64(restitution*cp.relativeVelocity)))
 
 			// clamp the accumulated impulse
 			newImpulse := maxFloat(cp.normalImpulse+impulse, 0.0)
